feat: Add Python script to check the value of the last digit
This commit adds a new Python script in the `python-if_else_loops_functions` directory. The script generates a random number and checks the value of its last digit. It then prints a message based on the value of the last digit. This feature enhances the functionality of the project by providing a way to analyze the last digit of a number.